PROMOTION WORKS
Lessons learnt:
Unity restricts access to its API from non-main threads... sometimes? Maybe only in the editor and not in the final build? I moved the data reception code to the main threads and my issues in the editor player stopped.
Task threads apparently keep going even if the main thread is closed, such as when "stopping" "the game" in the unity editor. I believe this was why the editor kept freezing and needing to be restarted. Lesson learned: Make your threads kill themselves.
Unity's Start() method doesn't run until the first update after an object begins to exist. This meant the method "GoHome" that effectively updates a piece's visual position by changing its GameObject's transform would throw an exception due to not having a reference to its renderer when changing its sprite's render order to reflect its new position. I made the method check if the renderer reference was null and remedy that if it was.