Added settings (don't work) and a matchmaking interface
A menu now pops up when matchmaking and there is a button to cancel it. I also added a very nice settings menu - I even intend to make it work!

Messed around with some tasks and waiting to make sure nothing is done in an appropriate order (saving tasks in variables).

The next step will be to add an account system. I've already done it for a PHP project so it shouldn't be very difficult, but it'll be quite a chore, especially when it comes to the dragging-and-dropping in unity aspect. A lot of people would probably enjoy that part more than coding, but I despise it. I suppose this is why it's a good idea to work together.